# swagger/openapi.yaml
openapi: 3.0.3
info:
  title: FMS API
  version: 1.0.0
  description: |
    Fleet Management System API - versi awal.
    Fitur utama:
    - Manajemen kendaraan
    - Posisi terkini kendaraan
    - Trip history & alert (read-only dulu)
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /vehicles:
    get:
      summary: Daftar kendaraan
      tags: [Vehicles]
      parameters:
        - in: query
          name: organizationId
          schema:
            type: integer
          required: false
          description: Filter berdasarkan organizationId
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
          required: false
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          required: false
      responses:
        '200':
          description: List kendaraan
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Vehicle'
                  total:
                    type: integer
    post:
      summary: Tambah kendaraan baru
      tags: [Vehicles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleCreateRequest'
      responses:
        '201':
          description: Kendaraan berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          description: Request tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /vehicles/{id}:
    get:
      summary: Detail kendaraan
      tags: [Vehicles]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detail kendaraan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '404':
          description: Kendaraan tidak ditemukan
    put:
      summary: Update data kendaraan
      tags: [Vehicles]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleUpdateRequest'
      responses:
        '200':
          description: Kendaraan berhasil diupdate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          description: Request tidak valid
        '404':
          description: Kendaraan tidak ditemukan
    delete:
      summary: Hapus/nonaktifkan kendaraan
      description: >
        Secara logis menonaktifkan kendaraan (soft delete) dengan mengubah `active = false`.
      tags: [Vehicles]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Berhasil dihapus/nonaktifkan
        '404':
          description: Kendaraan tidak ditemukan
  /vehicles/{id}/current-position:
    get:
      summary: Posisi terkini kendaraan
      tags: [Tracking]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Posisi terkini
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleCurrentPosition'
        '404':
          description: Data tidak ditemukan
...
components:
  schemas:
...
    Vehicle:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
          nullable: true
        plateNumber:
          type: string
          example: "B 1234 CD"
        name:
          type: string
          example: "Truk Box 01"
        vin:
          type: string
          nullable: true
        vehicleType:
          type: string
          example: "truck"
        active:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    VehicleCreateRequest:
      type: object
      required:
        - plateNumber
      properties:
        organizationId:
          type: integer
          nullable: true
        plateNumber:
          type: string
        name:
          type: string
        vin:
          type: string
        vehicleType:
          type: string
        metadata:
          type: object
          additionalProperties: true
    VehicleUpdateRequest:
      type: object
      properties:
        plateNumber:
          type: string
        name:
          type: string
        vin:
          type: string
        vehicleType:
          type: string
        active:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
    VehicleCurrentPosition:
      type: object
      properties:
        vehicleId:
          type: integer
        deviceId:
          type: integer
          nullable: true
        ts:
          type: string
          format: date-time
        lat:
          type: number
          format: double
          example: -6.200000
        lon:
          type: number
          format: double
          example: 106.816666
        speedKph:
          type: number
          format: float
          nullable: true
        headingDeg:
          type: number
          format: float
          nullable: true
        ignitionOn:
          type: boolean
          nullable: true
        odometerKm:
          type: number
          format: float
          nullable: true
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "bad_request"
        message:
          type: string
          example: "Plate number is required"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT